ai-skydeck-hackathon-new/django/requirements.txt
beautifulsoup4==4.12.3
djangorestframework==3.15.2
Requests==2.32.3
django-cors-headers==4.4.0
boto3